samples/multimodal-claude-thinking/requirements.txt
requests
boto3
PyPDF2
pillow
numpy 
matplotlib
scipy